+++ packages/components/ng-ovh-responsive-popover/rollup.config.js
@@ -1,0 +1,16 @@
+import rollupConfig from '@ovh-ux/component-rollup-config';
+
+const config = rollupConfig({
+  input: 'src/index.js',
+});
+
+export default [
+  config.cjs(),
+  config.umd({
+    output: {
+      globals: {
+        angular: 'angular',
+      },
+    },
+  }),
+];