+++ packages/components/ng-ovh-responsive-popover/src/ovh-angular-responsive-popover.mock.js
@@ -1,9 +1,0 @@
-"use strict";
-
-angular.module("responsivePopoverMock", ["ovh-angular-responsive-popover", "templates"]);
-
-angular.module("responsivePopoverMock").run(function ($httpBackend) {
-
-    $httpBackend.whenGET("my-test.html").respond(200, {});
-
-});